chore: change folder structure for nav component - write unit tests for contribute menu
