+++ web/src/components/layout/nav/contribute-menu.test.tsx
@@ -1,0 +1,36 @@
+import * as React from 'react';
+import { fireEvent, screen } from '@testing-library/react';
+
+import { ContributeMenu } from './contribute-menu';
+import { renderWithProviders } from '../../../../test/render-with-providers';
+
+describe('Contribute Menu Test', () => {
+  it('renders Contribute Menu', () => {
+    renderWithProviders(
+      <ContributeMenu
+        showMenu={false}
+        setShowMenu={jest.fn()}
+        showMobileMenu={false}
+        toggleMobileMenuVisible={jest.fn()}
+      />
+    );
+
+    expect(screen.queryByTestId('contribute-menu')).toBeTruthy();
+  });
+
+  it('shows the menu on hover', () => {
+    const setShowMenuSpy = jest.fn();
+
+    renderWithProviders(
+      <ContributeMenu
+        showMenu={false}
+        setShowMenu={setShowMenuSpy}
+        showMobileMenu={false}
+        toggleMobileMenuVisible={jest.fn()}
+      />
+    );
+
+    fireEvent.mouseEnter(screen.getByTestId('contribute-menu'));
+    expect(setShowMenuSpy).toHaveBeenCalledWith(true);
+  });
+});
